save model with files debug

+++ server/controllers/controller.js
@@ -2,9 +2,17 @@
  * Я не знаю почему не получается сделать через class, ошибка в this выходит
  * */
 
+const ApiError = require("../exceptions/ApiError");
+
+
 module.exports = (service) => ({
+
     async createOne(req, res, next){
         try{
+            // Предполагается что, по умолчанию доступ к api доступен только для аутентифицированных пользователей
+            if(!req.isAuthenticated()){
+                throw ApiError.UnauthorizedError('Unauthorized');
+            }
             const data = await service.createOne(req.body, req.files, req.user);
             res.json(data);
         }
@@ -15,6 +23,10 @@ module.exports = (service) => ({
 
     async findByQueryParams(req, res, next){
         try{
+            // Предполагается что, по умолчанию доступ к api доступен только для аутентифицированных пользователей
+            if(!req.isAuthenticated()){
+                throw ApiError.UnauthorizedError('Unauthorized');
+            }
             const data = await service.find(req.query, req.user);
             res.json(data);
         }
@@ -25,6 +37,10 @@ module.exports = (service) => ({
 
     async updateOne(req, res, next){
         try{
+            // Предполагается что, по умолчанию доступ к api доступен только для аутентифицированных пользователей
+            if(!req.isAuthenticated()){
+                throw ApiError.UnauthorizedError('Unauthorized');
+            }
             const data = await service.updateOne(req.body, req.files, req.user);
             res.json(data);
         }
@@ -35,6 +51,10 @@ module.exports = (service) => ({
 
     async deleteOne(req, res, next){
         try{
+            // Предполагается что, по умолчанию доступ к api доступен только для аутентифицированных пользователей
+            if(!req.isAuthenticated()){
+                throw ApiError.UnauthorizedError('Unauthorized');
+            }
             const data = await service.deleteOne(req.body, req.user);
             res.json(data);
         }
@@ -42,5 +62,6 @@ module.exports = (service) => ({
             next(err);
         }
     },
+
 });
